requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
cryptography==38.0.1
Deprecated==1.2.13
dj-database-url==1.0.0
Django==4.1.2
django-bootstrap5==22.1
docutils==0.19
flwave==1.0
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.0.0
jaraco.classes==3.2.3
jeepney==0.8.0
joblib==1.2.0
keyring==23.9.3
more-itertools==9.0.0
nltk==3.7
packaging==21.3
pdfminer.six==20220524
pdfplumber==0.7.5
Pillow==9.2.0
pkginfo==1.8.3
pycparser==2.21
Pygments==2.13.0
pyparsing==3.0.9
PyPDF2==2.11.1
python-dotenv==0.21.0
readme-renderer==37.2
redis==4.3.4
regex==2022.9.13
requests==2.28.1
requests-toolbelt==0.10.0
rfc3986==2.0.0
rich==12.6.0
rq==1.11.1
SecretStorage==3.3.3
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tqdm==4.64.1
twine==4.0.1
urllib3==1.26.12
Wand==0.6.10
webencodings==0.5.1
whitenoise==6.2.0
wrapt==1.14.1
zipp==3.9.0
